Module API.Quotation

type quotation = language:Ast.Scope.language -> State.t -> Ast.Loc.t -> string -> Ast.Term.t

val set_default_quotation : ?descriptor:Setup.quotations_descriptor -> quotation -> unit

The default quotation {{code}}

val register_named_quotation : ?descriptor:Setup.quotations_descriptor -> name:string -> quotation -> Ast.Scope.language

Named quotation {{name:code}}

val elpi_language : Ast.Scope.language

The anti-quotation to lambda Prolog

val elpi : quotation

Like quotations but for identifiers that begin and end with * "`" or "'", e.g. `this` and 'that'. Useful if the object language * needs something that looks like a string but with a custom compilation * (e.g. CD.string like but with a case insensitive comparison)

val declare_backtick : ?descriptor:Setup.quotations_descriptor -> name:string -> quotation -> Ast.Scope.language

val declare_singlequote : ?descriptor:Setup.quotations_descriptor -> name:string -> quotation -> Ast.Scope.language

val new_quotations_descriptor : unit -> Setup.quotations_descriptor