Module API.RawQuery

This module lets one generate a query by providing a RawData.term directly

val compile_ast : Compile.program -> Ast.query -> ( State.t -> State.t ) -> Compile.query

with the possibility to update the state in which the query will run

val compile_term : Compile.program -> ( State.t -> State.t * Ast.Term.t ) -> Compile.query

generate the query ast term with a function. The resulting term is typed, spilled, etc

val compile_raw_term : Compile.program -> ( State.t -> State.t * Data.term * Conversion.extra_goals ) -> Compile.query

generate the query term by hand, the result is used as is

val term_to_raw_term : State.t -> Compile.program -> ?ctx:RawData.constant Ast.Scope.Map.t -> depth:int -> Ast.Term.t -> State.t * Data.term

typechecks only if ctx is empty

val global_name_to_constant : State.t -> string -> RawData.constant

raises Not_found