Module Elpi_runtime.Data

module Fmt = Stdlib.Format

module F = Elpi_parser.Ast.Func

type indexing =

| MapOn of int

| Hash of int list

| DiscriminationTree of int list

Used to index the parameters of a predicate P

MapOn N -> N-th argument at depth 1 (head symbol only)

Hash L -> L is the list of depths given by the urer for the parameters of P. Indexing is done by hashing all the parameters with a non zero depth and comparing it with the hashing of the parameters of the query

DiscriminationTree L -> we use the same logic of Hash, except we use DiscriminationTree to discriminate clauses

val pp_indexing : Ppx_deriving_runtime.Format.formatter -> indexing -> Ppx_deriving_runtime.unit

val show_indexing : indexing -> Ppx_deriving_runtime.string

val compare_indexing : indexing -> indexing -> Ppx_deriving_runtime.int

type overlap_clause = {

overlap_loc : Elpi_util.Util.Loc.t option;

has_cut : bool;

mutable timestamp : int list;

arg_nb : int;

}

val pp_overlap_clause : Ppx_deriving_runtime.Format.formatter -> overlap_clause -> Ppx_deriving_runtime.unit

val show_overlap_clause : overlap_clause -> Ppx_deriving_runtime.string

type overlap =

| Allowed

| Forbidden of overlap_clause Discrimination_tree.t

val pp_overlap : Ppx_deriving_runtime.Format.formatter -> overlap -> Ppx_deriving_runtime.unit

val show_overlap : overlap -> Ppx_deriving_runtime.string

val mk_Forbidden : indexing -> overlap

type pred_info = {

indexing : indexing;

mode : Elpi_util.Util.Mode.hos;

overlap : overlap;

has_local_without_cut : Elpi_util.Util.Loc.t option;

}

val pp_pred_info : Ppx_deriving_runtime.Format.formatter -> pred_info -> Ppx_deriving_runtime.unit

val show_pred_info : pred_info -> Ppx_deriving_runtime.string

val same_indexing : pred_info -> pred_info -> bool

module Term : sig ... end

include module type of struct include Term end

val pp_oref : (Elpi_util.Util.UUID.t * Stdlib.Obj.t) Elpi_util.Util.spaghetti_printer

val id_term : Elpi_util.Util.UUID.t

type 'unification_def stuck_goal_kind = 'unification_def Term.stuck_goal_kind = ..

val pp_stuck_goal_kind : 'a -> 'b -> 'c -> unit

val show_stuck_goal_kind : 'a -> 'b -> string

val equal_stuck_goal_kind : 'a -> 'b -> 'b -> bool

type stuck_goal_kind +=

| Unification of 'unification_def

type ttype = Term.ttype =

| TConst of Elpi_util.Util.constant

| TApp of Elpi_util.Util.constant * ttype * ttype list

| TPred of bool * (Elpi_util.Util.Mode.t * ttype) list

| TArr of ttype * ttype

| TCData of Elpi_util.Util.CData.t

| TLam of ttype

val pp_ttype : Ppx_deriving_runtime.Format.formatter -> ttype -> Ppx_deriving_runtime.unit

val show_ttype : ttype -> Ppx_deriving_runtime.string

val compare_ttype : ttype -> ttype -> Ppx_deriving_runtime.int

type builtin_predicate = Term.builtin_predicate =

| Cut

| And

| Impl

| ImplBang

| RImpl

| Pi

| Sigma

| Eq

| Match

| Findall

| Delay

| Host of Elpi_util.Util.constant

val compare_builtin_predicate : builtin_predicate -> builtin_predicate -> Ppx_deriving_runtime.int

val pp_builtin_predicate : Ppx_deriving_runtime.Format.formatter -> builtin_predicate -> Ppx_deriving_runtime.unit

val builtin_predicates : builtin_predicate list

val builtin_predicate_max : int

val func_of_builtin_predicate : (Elpi_util.Util.constant -> F.t) -> builtin_predicate -> F.t

val show_builtin_predicate : ?table:'a -> (?table:'a -> Elpi_util.Util.constant -> string) -> builtin_predicate -> string

val const_of_builtin_predicate : builtin_predicate -> Elpi_util.Util.constant

val is_builtin_predicate : int -> bool

val builtin_predicate_of_const : int -> builtin_predicate

val map_builtin_predicate : (Elpi_util.Util.constant -> Elpi_util.Util.constant) -> builtin_predicate -> builtin_predicate

type term = Term.term =

| Const of Elpi_util.Util.constant

| Lam of term

| App of Elpi_util.Util.constant * term * term list

| Cons of term * term

| Nil

| Discard

| Builtin of builtin_predicate * term list

| CData of Elpi_util.Util.CData.t

| UVar of uvar * int

| AppUVar of uvar * term list

| Arg of int * int

| AppArg of int * term list

and uvar = Term.uvar = {

vardepth : int;

mutable contents : term;

mutable uid_private : int;

}

val pp_term : Ppx_deriving_runtime.Format.formatter -> term -> Ppx_deriving_runtime.unit

val show_term : term -> Ppx_deriving_runtime.string

val pp_uvar : Ppx_deriving_runtime.Format.formatter -> uvar -> Ppx_deriving_runtime.unit

val show_uvar : uvar -> Ppx_deriving_runtime.string

val compare_term : term -> term -> Ppx_deriving_runtime.int

val compare_uvar : uvar -> uvar -> Ppx_deriving_runtime.int

val cons2tcons : ?loc:Elpi_util.Util.Loc.t -> term -> ttype

val uvar_id : uvar -> int

val uvar_is_a_blocker : uvar -> bool

val uvar_isnt_a_blocker : uvar -> bool

val uvar_set_blocker : uvar -> unit

val uvar_unset_blocker : uvar -> unit

type clause = Term.clause = {

depth : int;

args : term list;

hyps : term list;

vars : int;

mode : Elpi_util.Util.Mode.hos;

loc : Elpi_util.Util.Loc.t option;

mutable timestamp : int list;

}

val pp_clause : Ppx_deriving_runtime.Format.formatter -> clause -> Ppx_deriving_runtime.unit

val show_clause : clause -> Ppx_deriving_runtime.string

val compare_clause : clause -> clause -> Ppx_deriving_runtime.int

type grafting_time = int list

val pp_grafting_time : Ppx_deriving_runtime.Format.formatter -> grafting_time -> Ppx_deriving_runtime.unit

val show_grafting_time : grafting_time -> Ppx_deriving_runtime.string

val compare_grafting_time : grafting_time -> grafting_time -> Ppx_deriving_runtime.int

val compare_constant : Elpi_util.Util.constant -> Elpi_util.Util.constant -> int

type times = (grafting_time * Elpi_util.Util.constant) Elpi_util.Util.StrMap.t

val pp_times : Ppx_deriving_runtime.Format.formatter -> times -> Ppx_deriving_runtime.unit

val show_times : times -> Ppx_deriving_runtime.string

val compare_times : times -> times -> Ppx_deriving_runtime.int

type stuck_goal = Term.stuck_goal = {

mutable blockers : blockers;

kind : unification_def stuck_goal_kind;

}

and blockers = uvar list

and unification_def = Term.unification_def = {

adepth : int;

env : term array;

bdepth : int;

a : term;

b : term;

matching : bool;

}

and clause_src = Term.clause_src = {

hdepth : int;

hsrc : term;

}

and prolog_prog = Term.prolog_prog = {

src : clause_src list;

index : index;

}

and clause_list = clause Bl.t

and index = first_lvl_idx

and first_lvl_idx = Term.first_lvl_idx = {

idx : second_lvl_idx Ptmap.t;

time : int;

times : times;

}

and second_lvl_idx = Term.second_lvl_idx =

| TwoLevelIndex of {

mode : Elpi_util.Util.Mode.hos;

argno : int;

all_clauses : clause_list;

flex_arg_clauses : clause_list;

arg_idx : clause_list Ptmap.t;

}

| BitHash of {

mode : Elpi_util.Util.Mode.hos;

args : int list;

args_idx : clause_list Ptmap.t;

}

| IndexWithDiscriminationTree of {

mode : Elpi_util.Util.Mode.hos;

arg_depths : int list;

args_idx : clause Discrimination_tree.t;

}

val pp_stuck_goal : Ppx_deriving_runtime.Format.formatter -> stuck_goal -> Ppx_deriving_runtime.unit

val show_stuck_goal : stuck_goal -> Ppx_deriving_runtime.string

val pp_blockers : Ppx_deriving_runtime.Format.formatter -> blockers -> Ppx_deriving_runtime.unit

val show_blockers : blockers -> Ppx_deriving_runtime.string

val pp_unification_def : Ppx_deriving_runtime.Format.formatter -> unification_def -> Ppx_deriving_runtime.unit

val show_unification_def : unification_def -> Ppx_deriving_runtime.string

val pp_clause_src : Ppx_deriving_runtime.Format.formatter -> clause_src -> Ppx_deriving_runtime.unit

val show_clause_src : clause_src -> Ppx_deriving_runtime.string

val pp_prolog_prog : Ppx_deriving_runtime.Format.formatter -> prolog_prog -> Ppx_deriving_runtime.unit

val show_prolog_prog : prolog_prog -> Ppx_deriving_runtime.string

val pp_clause_list : Ppx_deriving_runtime.Format.formatter -> clause_list -> Ppx_deriving_runtime.unit

val show_clause_list : clause_list -> Ppx_deriving_runtime.string

val pp_index : Ppx_deriving_runtime.Format.formatter -> index -> Ppx_deriving_runtime.unit

val show_index : index -> Ppx_deriving_runtime.string

val pp_first_lvl_idx : Ppx_deriving_runtime.Format.formatter -> index -> Ppx_deriving_runtime.unit

val show_first_lvl_idx : index -> Ppx_deriving_runtime.string

val pp_second_lvl_idx : Ppx_deriving_runtime.Format.formatter -> second_lvl_idx -> Ppx_deriving_runtime.unit

val show_second_lvl_idx : second_lvl_idx -> Ppx_deriving_runtime.string

val stop : bool Stdlib.ref

val close_index : index -> index

type constraints = stuck_goal list

type hyps = clause_src list

type suspended_goal = Term.suspended_goal = {

context : hyps;

goal : int * term;

blockers : blockers;

}

val mkLam : term -> term

val mkApp : Elpi_util.Util.constant -> term -> term list -> term

val mkCons : term -> term -> term

val mkNil : term

val mkDiscard : term

val mkBuiltin : builtin_predicate -> term list -> term

val mkCData : Elpi_util.Util.CData.t -> term

val mkUVar : uvar -> int -> term

val mkAppUVar : uvar -> term list -> term

val mkArg : int -> int -> term

val mkAppArg : int -> term list -> term

module C = Term.C

val destConst : term -> Elpi_util.Util.constant

val oref : depth:int -> term -> uvar

val (!!) : uvar -> term

type env = term array

val empty_env : 'a array

module State : sig ... end

val elpi_state_descriptor : State.descriptor

type core_symbol =

| As

| Uv

| ECons

| ENil

val min_core_symbol : int

val max_core_symbol : int

val core_symbol_to_enum : core_symbol -> int

val core_symbol_of_enum : int -> core_symbol option

val compare_core_symbol : core_symbol -> core_symbol -> Ppx_deriving_runtime.int

val pp_core_symbol : Ppx_deriving_runtime.Format.formatter -> core_symbol -> Ppx_deriving_runtime.unit

val show_core_symbol : core_symbol -> Ppx_deriving_runtime.string

val func_of_core_symbol : core_symbol -> F.t

val is_core_symbol : F.t -> bool

module Symbol : sig ... end

module Global_symbols : sig ... end

val dummy : term

val dummy_uvar_body : uvar

module CHR : sig ... end

type clause_w_info = {

clloc : Elpi_util.Util.CData.t;

clargsname : string list;

clbody : clause;

}

val pp_clause_w_info : Ppx_deriving_runtime.Format.formatter -> clause_w_info -> Ppx_deriving_runtime.unit

val show_clause_w_info : clause_w_info -> Ppx_deriving_runtime.string

exception No_clause

exception No_more_steps

exception Flex_head

module Conversion : sig ... end

module ContextualConversion : sig ... end

val while_compiling : bool State.component

module HoasHooks : sig ... end

module CalcHooks : sig ... end

module BuiltInPredicate : sig ... end

type symbol_table = {

mutable c2s : Symbol.t Elpi_util.Util.Constants.Map.t;

c2t : (Elpi_util.Util.constant, term) Elpi_util.Util.Hashtbl.t;

mutable frozen_constants : int;

}

val pp_symbol_table : Ppx_deriving_runtime.Format.formatter -> symbol_table -> Ppx_deriving_runtime.unit

val show_symbol_table : symbol_table -> Ppx_deriving_runtime.string

type executable = {

compiled_program : prolog_prog;

chr : CHR.t;

initial_depth : int;

initial_goal : term;

initial_runtime_state : State.t;

symbol_table : symbol_table;

builtins : BuiltInPredicate.builtin_table;

assignments : term Elpi_util.Util.StrMap.t;

}

type pp_ctx = {

uv_names : (string Elpi_util.Util.IntMap.t * int) Stdlib.ref;

table : symbol_table;

}

type solution = {

assignments : term Elpi_util.Util.StrMap.t;

constraints : constraints;

state : State.t;

pp_ctx : pp_ctx;

state_for_relocation : int * symbol_table;

}

type 'a outcome =

| Success of solution

| Failure

| NoMoreSteps

exception CannotDeclareClauseForBuiltin of Elpi_util.Util.Loc.t option * builtin_predicate