Module Data.Term

val pp_oref : (Elpi_util.Util.UUID.t * Stdlib.Obj.t) Elpi_util.Util.spaghetti_printer

val id_term : Elpi_util.Util.UUID.t

type 'unification_def stuck_goal_kind = ..

val pp_stuck_goal_kind : 'a -> 'b -> 'c -> unit

val show_stuck_goal_kind : 'a -> 'b -> string

val equal_stuck_goal_kind : 'a -> 'b -> 'b -> bool

type stuck_goal_kind +=

| Unification of 'unification_def

type arg_mode = Elpi_util.Util.arg_mode =

| Input

| Output

val pp_arg_mode : Ppx_deriving_runtime.Format.formatter -> arg_mode -> Ppx_deriving_runtime.unit

val show_arg_mode : arg_mode -> Ppx_deriving_runtime.string

val compare_arg_mode : arg_mode -> arg_mode -> Ppx_deriving_runtime.int

type mode_aux = Elpi_util.Util.mode_aux =

| Fo of arg_mode

| Ho of arg_mode * mode

and mode = mode_aux list

val pp_mode_aux : Ppx_deriving_runtime.Format.formatter -> mode_aux -> Ppx_deriving_runtime.unit

val show_mode_aux : mode_aux -> Ppx_deriving_runtime.string

val pp_mode : Ppx_deriving_runtime.Format.formatter -> mode -> Ppx_deriving_runtime.unit

val show_mode : mode -> Ppx_deriving_runtime.string

val compare_mode_aux : mode_aux -> mode_aux -> Ppx_deriving_runtime.int

val compare_mode : mode -> mode -> Ppx_deriving_runtime.int

type ttype =

| TConst of Elpi_util.Util.constant

| TApp of Elpi_util.Util.constant * ttype * ttype list

| TPred of bool * (arg_mode * ttype) list

| TArr of ttype * ttype

| TCData of Elpi_util.Util.CData.t

| TLam of ttype

val pp_ttype : Ppx_deriving_runtime.Format.formatter -> ttype -> Ppx_deriving_runtime.unit

val show_ttype : ttype -> Ppx_deriving_runtime.string

val compare_ttype : ttype -> ttype -> Ppx_deriving_runtime.int

type term =

| Const of Elpi_util.Util.constant

| Lam of term

| App of Elpi_util.Util.constant * term * term list

| Cons of term * term

| Nil

| Discard

| Builtin of Elpi_util.Util.constant * term list

| CData of Elpi_util.Util.CData.t

| UVar of uvar_body * int * int

| AppUVar of uvar_body * int * term list

| Arg of int * int

| AppArg of int * term list

and uvar_body = {

mutable contents : term;

mutable uid_private : int;

}

val pp_term : Ppx_deriving_runtime.Format.formatter -> term -> Ppx_deriving_runtime.unit

val show_term : term -> Ppx_deriving_runtime.string

val pp_uvar_body : Ppx_deriving_runtime.Format.formatter -> uvar_body -> Ppx_deriving_runtime.unit

val show_uvar_body : uvar_body -> Ppx_deriving_runtime.string

val compare_term : term -> term -> Ppx_deriving_runtime.int

val compare_uvar_body : uvar_body -> uvar_body -> Ppx_deriving_runtime.int

val cons2tcons : ?loc:Elpi_util.Util.Loc.t -> term -> ttype

val uvar_id : uvar_body -> int

val uvar_is_a_blocker : uvar_body -> bool

val uvar_isnt_a_blocker : uvar_body -> bool

val uvar_set_blocker : uvar_body -> unit

val uvar_unset_blocker : uvar_body -> unit

type clause = {

depth : int;

args : term list;

hyps : term list;

vars : int;

mode : mode;

loc : Elpi_util.Util.Loc.t option;

mutable timestamp : int list;

}

val pp_clause : Ppx_deriving_runtime.Format.formatter -> clause -> Ppx_deriving_runtime.unit

val show_clause : clause -> Ppx_deriving_runtime.string

val compare_clause : clause -> clause -> Ppx_deriving_runtime.int

val get_arg_mode : mode_aux -> arg_mode

val to_mode : bool -> mode_aux

type grafting_time = int list

val pp_grafting_time : Ppx_deriving_runtime.Format.formatter -> grafting_time -> Ppx_deriving_runtime.unit

val show_grafting_time : grafting_time -> Ppx_deriving_runtime.string

val compare_grafting_time : grafting_time -> grafting_time -> Ppx_deriving_runtime.int

val compare_constant : Elpi_util.Util.constant -> Elpi_util.Util.constant -> int

type times = (grafting_time * Elpi_util.Util.constant) Elpi_util.Util.StrMap.t

val pp_times : Ppx_deriving_runtime.Format.formatter -> times -> Ppx_deriving_runtime.unit

val show_times : times -> Ppx_deriving_runtime.string

val compare_times : times -> times -> Ppx_deriving_runtime.int

type stuck_goal = {

mutable blockers : blockers;

kind : unification_def stuck_goal_kind;

}

and blockers = uvar_body list

and unification_def = {

adepth : int;

env : term array;

bdepth : int;

a : term;

b : term;

matching : bool;

}

and clause_src = {

hdepth : int;

hsrc : term;

}

and prolog_prog = {

src : clause_src list;

index : index;

}

and clause_list = clause Bl.t

and index = first_lvl_idx

and first_lvl_idx = {

idx : second_lvl_idx Ptmap.t;

time : int;

times : times;

}

and second_lvl_idx =

| TwoLevelIndex of {

mode : mode;

argno : int;

all_clauses : clause_list;

flex_arg_clauses : clause_list;

arg_idx : clause_list Ptmap.t;

}

| BitHash of {

mode : mode;

args : int list;

args_idx : clause_list Ptmap.t;

}

| IndexWithDiscriminationTree of {

mode : mode;

arg_depths : int list;

args_idx : clause Discrimination_tree.t;

}

val pp_stuck_goal : Ppx_deriving_runtime.Format.formatter -> stuck_goal -> Ppx_deriving_runtime.unit

val show_stuck_goal : stuck_goal -> Ppx_deriving_runtime.string

val pp_blockers : Ppx_deriving_runtime.Format.formatter -> blockers -> Ppx_deriving_runtime.unit

val show_blockers : blockers -> Ppx_deriving_runtime.string

val pp_unification_def : Ppx_deriving_runtime.Format.formatter -> unification_def -> Ppx_deriving_runtime.unit

val show_unification_def : unification_def -> Ppx_deriving_runtime.string

val pp_clause_src : Ppx_deriving_runtime.Format.formatter -> clause_src -> Ppx_deriving_runtime.unit

val show_clause_src : clause_src -> Ppx_deriving_runtime.string

val pp_prolog_prog : Ppx_deriving_runtime.Format.formatter -> prolog_prog -> Ppx_deriving_runtime.unit

val show_prolog_prog : prolog_prog -> Ppx_deriving_runtime.string

val pp_clause_list : Ppx_deriving_runtime.Format.formatter -> clause_list -> Ppx_deriving_runtime.unit

val show_clause_list : clause_list -> Ppx_deriving_runtime.string

val pp_index : Ppx_deriving_runtime.Format.formatter -> index -> Ppx_deriving_runtime.unit

val show_index : index -> Ppx_deriving_runtime.string

val pp_first_lvl_idx : Ppx_deriving_runtime.Format.formatter -> index -> Ppx_deriving_runtime.unit

val show_first_lvl_idx : index -> Ppx_deriving_runtime.string

val pp_second_lvl_idx : Ppx_deriving_runtime.Format.formatter -> second_lvl_idx -> Ppx_deriving_runtime.unit

val show_second_lvl_idx : second_lvl_idx -> Ppx_deriving_runtime.string

val stop : bool Stdlib.ref

val close_index : index -> index

type constraints = stuck_goal list

type hyps = clause_src list

type suspended_goal = {

context : hyps;

goal : int * term;

}

type indexing =

| MapOn of int

| Hash of int list

| DiscriminationTree of int list

Used to index the parameters of a predicate P

MapOn N -> N-th argument at depth 1 (head symbol only)

Hash L -> L is the list of depths given by the urer for the parameters of P. Indexing is done by hashing all the parameters with a non zero depth and comparing it with the hashing of the parameters of the query

DiscriminationTree L -> we use the same logic of Hash, except we use DiscriminationTree to discriminate clauses

val pp_indexing : Ppx_deriving_runtime.Format.formatter -> indexing -> Ppx_deriving_runtime.unit

val show_indexing : indexing -> Ppx_deriving_runtime.string

val mkLam : term -> term

val mkApp : Elpi_util.Util.constant -> term -> term list -> term

val mkCons : term -> term -> term

val mkNil : term

val mkDiscard : term

val mkBuiltin : Elpi_util.Util.constant -> term list -> term

val mkCData : Elpi_util.Util.CData.t -> term

val mkUVar : uvar_body -> int -> int -> term

val mkAppUVar : uvar_body -> int -> term list -> term

val mkArg : int -> int -> term

val mkAppArg : int -> term list -> term

module C : sig ... end

val destConst : term -> Elpi_util.Util.constant

val oref : term -> uvar_body

val (!!) : uvar_body -> term

type env = term array

val empty_env : 'a array