Module Elpi_trace_atd.Trace_atd

type builtin_kind = [

| `Logic

| `FFI

]

type builtin_rule = {

name : string;

kind : builtin_kind;

payload : string list;

}

type chr_text = string

type color = [

| `Red

| `Green

| `Grey

| `YellowGreen

| `YellowRed

]

type file_location = {

filename : string;

line : int;

column : int;

character : int;

}

type goal_id = int

type event = [

| `Assign of string

| `Fail of string

| `ResumeGoal of goal_id list

]

type goal = {

goal_text : string;

goal_id : goal_id;

}

type goal_text = string

type outcome = [

| `Fail

| `Success

]

type resume = goal list

type rule_text = string

type runtime_id = int

type solution_text = string list

type step_id = int

type location = [

| `File of file_location

| `Context of step_id

]

type user_rule = {

rule_text : rule_text;

rule_loc : location;

}

type cutted = {

cut_branch_for_goal : goal;

cut_branch : user_rule;

}

type cut = {

cut_goal_id : goal_id;

cut_victims : cutted list;

}

type rule = [

| `BuiltinRule of builtin_rule

| `UserRule of user_rule

]

type attempt = {

rule : rule;

events : event list;

}

type frame = {

rule : rule;

step_id : step_id;

runtime_id : runtime_id;

}

type stack = frame list

type successful_attempt = {

attempt : attempt;

siblings : goal list;

siblings_aggregated_outcome : outcome;

}

type inference = {

current_goal_id : goal_id;

current_goal_text : goal_text;

current_goal_predicate : string;

failed_attempts : attempt list;

successful_attempts : successful_attempt list;

more_successful_attempts : step_id list;

more_failing_attempts : step_id list;

stack : stack;

}

type suspend = {

suspend_goal_id : goal_id;

suspend_goal_text : goal_text;

suspend_sibling : goal;

suspend_stack : stack;

}

type card = {

step_id : step_id;

runtime_id : runtime_id;

step : step;

color : color;

}

and chr = {

chr_failed_attempts : chr_attempt list;

chr_successful_attempts : successful_chr_attempt list;

chr_store_before : goal list;

chr_store_after : goal list;

}

and chr_attempt = {

chr_text : chr_text;

chr_loc : file_location;

chr_condition_cards : trace;

}

and findall = {

findall_goal_id : goal_id;

findall_goal_text : goal_text;

findall_cards : trace;

findall_solution_text : solution_text;

findall_stack : stack;

}

and step = [

| `Inference of inference

| `Findall of findall

| `Cut of cut

| `Suspend of suspend

| `Resume of resume

| `CHR of chr

| `Init of goal

]

and successful_chr_attempt = {

chr_attempt : chr_attempt;

chr_removed_goals : goal_id list;

chr_new_goals : goal list;

}

and trace = card list

type stop = {

cause : string;

time : float;

}

type kind = [

| `Start

| `Info

| `Stop of stop

]

type item = {

step : int;

kind : kind list;

goal_id : int;

runtime_id : int;

name : string;

payload : string list;

}

type raw_trace = item list

val write_builtin_kind : Stdlib.Buffer.t -> builtin_kind -> unit

Output a JSON value of type builtin_kind.

val string_of_builtin_kind : ?len:int -> builtin_kind -> string

Serialize a value of type builtin_kind into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_builtin_kind : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> builtin_kind

Input JSON data of type builtin_kind.

val builtin_kind_of_string : string -> builtin_kind

Deserialize JSON data of type builtin_kind.

val write_builtin_rule : Stdlib.Buffer.t -> builtin_rule -> unit

Output a JSON value of type builtin_rule.

val string_of_builtin_rule : ?len:int -> builtin_rule -> string

Serialize a value of type builtin_rule into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_builtin_rule : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> builtin_rule

Input JSON data of type builtin_rule.

val builtin_rule_of_string : string -> builtin_rule

Deserialize JSON data of type builtin_rule.

val create_builtin_rule : name:string -> kind:builtin_kind -> payload:string list -> unit -> builtin_rule

Create a record of type builtin_rule.

val write_chr_text : Stdlib.Buffer.t -> chr_text -> unit

Output a JSON value of type chr_text.

val string_of_chr_text : ?len:int -> chr_text -> string

Serialize a value of type chr_text into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_chr_text : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> chr_text

Input JSON data of type chr_text.

val chr_text_of_string : string -> chr_text

Deserialize JSON data of type chr_text.

val write_color : Stdlib.Buffer.t -> color -> unit

Output a JSON value of type color.

val string_of_color : ?len:int -> color -> string

Serialize a value of type color into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_color : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> color

Input JSON data of type color.

val color_of_string : string -> color

Deserialize JSON data of type color.

val write_file_location : Stdlib.Buffer.t -> file_location -> unit

Output a JSON value of type file_location.

val string_of_file_location : ?len:int -> file_location -> string

Serialize a value of type file_location into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_file_location : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> file_location

Input JSON data of type file_location.

val file_location_of_string : string -> file_location

Deserialize JSON data of type file_location.

val create_file_location : filename:string -> line:int -> column:int -> character:int -> unit -> file_location

Create a record of type file_location.

val write_goal_id : Stdlib.Buffer.t -> goal_id -> unit

Output a JSON value of type goal_id.

val string_of_goal_id : ?len:int -> goal_id -> string

Serialize a value of type goal_id into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_goal_id : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> goal_id

Input JSON data of type goal_id.

val goal_id_of_string : string -> goal_id

Deserialize JSON data of type goal_id.

val write_event : Stdlib.Buffer.t -> event -> unit

Output a JSON value of type event.

val string_of_event : ?len:int -> event -> string

Serialize a value of type event into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_event : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> event

Input JSON data of type event.

val event_of_string : string -> event

Deserialize JSON data of type event.

val write_goal : Stdlib.Buffer.t -> goal -> unit

Output a JSON value of type goal.

val string_of_goal : ?len:int -> goal -> string

Serialize a value of type goal into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_goal : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> goal

Input JSON data of type goal.

val goal_of_string : string -> goal

Deserialize JSON data of type goal.

val create_goal : goal_text:string -> goal_id:goal_id -> unit -> goal

Create a record of type goal.

val write_goal_text : Stdlib.Buffer.t -> goal_text -> unit

Output a JSON value of type goal_text.

val string_of_goal_text : ?len:int -> goal_text -> string

Serialize a value of type goal_text into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_goal_text : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> goal_text

Input JSON data of type goal_text.

val goal_text_of_string : string -> goal_text

Deserialize JSON data of type goal_text.

val write_outcome : Stdlib.Buffer.t -> outcome -> unit

Output a JSON value of type outcome.

val string_of_outcome : ?len:int -> outcome -> string

Serialize a value of type outcome into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_outcome : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> outcome

Input JSON data of type outcome.

val outcome_of_string : string -> outcome

Deserialize JSON data of type outcome.

val write_resume : Stdlib.Buffer.t -> resume -> unit

Output a JSON value of type resume.

val string_of_resume : ?len:int -> resume -> string

Serialize a value of type resume into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_resume : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> resume

Input JSON data of type resume.

val resume_of_string : string -> resume

Deserialize JSON data of type resume.

val write_rule_text : Stdlib.Buffer.t -> rule_text -> unit

Output a JSON value of type rule_text.

val string_of_rule_text : ?len:int -> rule_text -> string

Serialize a value of type rule_text into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_rule_text : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> rule_text

Input JSON data of type rule_text.

val rule_text_of_string : string -> rule_text

Deserialize JSON data of type rule_text.

val write_runtime_id : Stdlib.Buffer.t -> runtime_id -> unit

Output a JSON value of type runtime_id.

val string_of_runtime_id : ?len:int -> runtime_id -> string

Serialize a value of type runtime_id into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_runtime_id : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> runtime_id

Input JSON data of type runtime_id.

val runtime_id_of_string : string -> runtime_id

Deserialize JSON data of type runtime_id.

val write_solution_text : Stdlib.Buffer.t -> solution_text -> unit

Output a JSON value of type solution_text.

val string_of_solution_text : ?len:int -> solution_text -> string

Serialize a value of type solution_text into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_solution_text : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> solution_text

Input JSON data of type solution_text.

val solution_text_of_string : string -> solution_text

Deserialize JSON data of type solution_text.

val write_step_id : Stdlib.Buffer.t -> step_id -> unit

Output a JSON value of type step_id.

val string_of_step_id : ?len:int -> step_id -> string

Serialize a value of type step_id into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_step_id : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> step_id

Input JSON data of type step_id.

val step_id_of_string : string -> step_id

Deserialize JSON data of type step_id.

val write_location : Stdlib.Buffer.t -> location -> unit

Output a JSON value of type location.

val string_of_location : ?len:int -> location -> string

Serialize a value of type location into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_location : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> location

Input JSON data of type location.

val location_of_string : string -> location

Deserialize JSON data of type location.

val write_user_rule : Stdlib.Buffer.t -> user_rule -> unit

Output a JSON value of type user_rule.

val string_of_user_rule : ?len:int -> user_rule -> string

Serialize a value of type user_rule into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_user_rule : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> user_rule

Input JSON data of type user_rule.

val user_rule_of_string : string -> user_rule

Deserialize JSON data of type user_rule.

val create_user_rule : rule_text:rule_text -> rule_loc:location -> unit -> user_rule

Create a record of type user_rule.

val write_cutted : Stdlib.Buffer.t -> cutted -> unit

Output a JSON value of type cutted.

val string_of_cutted : ?len:int -> cutted -> string

Serialize a value of type cutted into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_cutted : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> cutted

Input JSON data of type cutted.

val cutted_of_string : string -> cutted

Deserialize JSON data of type cutted.

val create_cutted : cut_branch_for_goal:goal -> cut_branch:user_rule -> unit -> cutted

Create a record of type cutted.

val write_cut : Stdlib.Buffer.t -> cut -> unit

Output a JSON value of type cut.

val string_of_cut : ?len:int -> cut -> string

Serialize a value of type cut into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_cut : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> cut

Input JSON data of type cut.

val cut_of_string : string -> cut

Deserialize JSON data of type cut.

val create_cut : cut_goal_id:goal_id -> cut_victims:cutted list -> unit -> cut

Create a record of type cut.

val write_rule : Stdlib.Buffer.t -> rule -> unit

Output a JSON value of type rule.

val string_of_rule : ?len:int -> rule -> string

Serialize a value of type rule into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_rule : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> rule

Input JSON data of type rule.

val rule_of_string : string -> rule

Deserialize JSON data of type rule.

val write_attempt : Stdlib.Buffer.t -> attempt -> unit

Output a JSON value of type attempt.

val string_of_attempt : ?len:int -> attempt -> string

Serialize a value of type attempt into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_attempt : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> attempt

Input JSON data of type attempt.

val attempt_of_string : string -> attempt

Deserialize JSON data of type attempt.

val create_attempt : rule:rule -> events:event list -> unit -> attempt

Create a record of type attempt.

val write_frame : Stdlib.Buffer.t -> frame -> unit

Output a JSON value of type frame.

val string_of_frame : ?len:int -> frame -> string

Serialize a value of type frame into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_frame : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> frame

Input JSON data of type frame.

val frame_of_string : string -> frame

Deserialize JSON data of type frame.

val create_frame : rule:rule -> step_id:step_id -> runtime_id:runtime_id -> unit -> frame

Create a record of type frame.

val write_stack : Stdlib.Buffer.t -> stack -> unit

Output a JSON value of type stack.

val string_of_stack : ?len:int -> stack -> string

Serialize a value of type stack into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_stack : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> stack

Input JSON data of type stack.

val stack_of_string : string -> stack

Deserialize JSON data of type stack.

val write_successful_attempt : Stdlib.Buffer.t -> successful_attempt -> unit

Output a JSON value of type successful_attempt.

val string_of_successful_attempt : ?len:int -> successful_attempt -> string

Serialize a value of type successful_attempt into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_successful_attempt : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> successful_attempt

Input JSON data of type successful_attempt.

val successful_attempt_of_string : string -> successful_attempt

Deserialize JSON data of type successful_attempt.

val create_successful_attempt : attempt:attempt -> siblings:goal list -> siblings_aggregated_outcome:outcome -> unit -> successful_attempt

Create a record of type successful_attempt.

val write_inference : Stdlib.Buffer.t -> inference -> unit

Output a JSON value of type inference.

val string_of_inference : ?len:int -> inference -> string

Serialize a value of type inference into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_inference : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> inference

Input JSON data of type inference.

val inference_of_string : string -> inference

Deserialize JSON data of type inference.

val create_inference : current_goal_id:goal_id -> current_goal_text:goal_text -> current_goal_predicate:string -> failed_attempts:attempt list -> successful_attempts:successful_attempt list -> more_successful_attempts:step_id list -> ?more_failing_attempts:step_id list -> stack:stack -> unit -> inference

Create a record of type inference.

val write_suspend : Stdlib.Buffer.t -> suspend -> unit

Output a JSON value of type suspend.

val string_of_suspend : ?len:int -> suspend -> string

Serialize a value of type suspend into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_suspend : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> suspend

Input JSON data of type suspend.

val suspend_of_string : string -> suspend

Deserialize JSON data of type suspend.

val create_suspend : suspend_goal_id:goal_id -> suspend_goal_text:goal_text -> suspend_sibling:goal -> suspend_stack:stack -> unit -> suspend

Create a record of type suspend.

val write_card : Stdlib.Buffer.t -> card -> unit

Output a JSON value of type card.

val string_of_card : ?len:int -> card -> string

Serialize a value of type card into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_card : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> card

Input JSON data of type card.

val card_of_string : string -> card

Deserialize JSON data of type card.

val create_card : step_id:step_id -> runtime_id:runtime_id -> step:step -> color:color -> unit -> card

Create a record of type card.

val write_chr : Stdlib.Buffer.t -> chr -> unit

Output a JSON value of type chr.

val string_of_chr : ?len:int -> chr -> string

Serialize a value of type chr into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_chr : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> chr

Input JSON data of type chr.

val chr_of_string : string -> chr

Deserialize JSON data of type chr.

val create_chr : chr_failed_attempts:chr_attempt list -> chr_successful_attempts:successful_chr_attempt list -> chr_store_before:goal list -> chr_store_after:goal list -> unit -> chr

Create a record of type chr.

val write_chr_attempt : Stdlib.Buffer.t -> chr_attempt -> unit

Output a JSON value of type chr_attempt.

val string_of_chr_attempt : ?len:int -> chr_attempt -> string

Serialize a value of type chr_attempt into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_chr_attempt : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> chr_attempt

Input JSON data of type chr_attempt.

val chr_attempt_of_string : string -> chr_attempt

Deserialize JSON data of type chr_attempt.

val create_chr_attempt : chr_text:chr_text -> chr_loc:file_location -> chr_condition_cards:trace -> unit -> chr_attempt

Create a record of type chr_attempt.

val write_findall : Stdlib.Buffer.t -> findall -> unit

Output a JSON value of type findall.

val string_of_findall : ?len:int -> findall -> string

Serialize a value of type findall into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_findall : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> findall

Input JSON data of type findall.

val findall_of_string : string -> findall

Deserialize JSON data of type findall.

val create_findall : findall_goal_id:goal_id -> findall_goal_text:goal_text -> findall_cards:trace -> findall_solution_text:solution_text -> findall_stack:stack -> unit -> findall

Create a record of type findall.

val write_step : Stdlib.Buffer.t -> step -> unit

Output a JSON value of type step.

val string_of_step : ?len:int -> step -> string

Serialize a value of type step into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_step : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> step

Input JSON data of type step.

val step_of_string : string -> step

Deserialize JSON data of type step.

val write_successful_chr_attempt : Stdlib.Buffer.t -> successful_chr_attempt -> unit

Output a JSON value of type successful_chr_attempt.

val string_of_successful_chr_attempt : ?len:int -> successful_chr_attempt -> string

Serialize a value of type successful_chr_attempt into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_successful_chr_attempt : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> successful_chr_attempt

Input JSON data of type successful_chr_attempt.

val successful_chr_attempt_of_string : string -> successful_chr_attempt

Deserialize JSON data of type successful_chr_attempt.

val create_successful_chr_attempt : chr_attempt:chr_attempt -> chr_removed_goals:goal_id list -> chr_new_goals:goal list -> unit -> successful_chr_attempt

Create a record of type successful_chr_attempt.

val write_trace : Stdlib.Buffer.t -> trace -> unit

Output a JSON value of type trace.

val string_of_trace : ?len:int -> trace -> string

Serialize a value of type trace into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_trace : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> trace

Input JSON data of type trace.

val trace_of_string : string -> trace

Deserialize JSON data of type trace.

val write_stop : Stdlib.Buffer.t -> stop -> unit

Output a JSON value of type stop.

val string_of_stop : ?len:int -> stop -> string

Serialize a value of type stop into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_stop : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> stop

Input JSON data of type stop.

val stop_of_string : string -> stop

Deserialize JSON data of type stop.

val create_stop : cause:string -> time:float -> unit -> stop

Create a record of type stop.

val write_kind : Stdlib.Buffer.t -> kind -> unit

Output a JSON value of type kind.

val string_of_kind : ?len:int -> kind -> string

Serialize a value of type kind into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_kind : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> kind

Input JSON data of type kind.

val kind_of_string : string -> kind

Deserialize JSON data of type kind.

val write_item : Stdlib.Buffer.t -> item -> unit

Output a JSON value of type item.

val string_of_item : ?len:int -> item -> string

Serialize a value of type item into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_item : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> item

Input JSON data of type item.

val item_of_string : string -> item

Deserialize JSON data of type item.

val create_item : step:int -> kind:kind list -> goal_id:int -> runtime_id:int -> name:string -> payload:string list -> unit -> item

Create a record of type item.

val write_raw_trace : Stdlib.Buffer.t -> raw_trace -> unit

Output a JSON value of type raw_trace.

val string_of_raw_trace : ?len:int -> raw_trace -> string

Serialize a value of type raw_trace into a JSON string.

parameter len
specifies the initial length of the buffer used internally. Default: 1024.

val read_raw_trace : Yojson.Safe.lexer_state -> Stdlib.Lexing.lexbuf -> raw_trace

Input JSON data of type raw_trace.

val raw_trace_of_string : string -> raw_trace

Deserialize JSON data of type raw_trace.