Module API.RawData

This module exposes the low level representation of terms. * * The data type term is opaque and can only be accessed by using the * look API that exposes a term view. The look view automatically * substitutes assigned unification variables by their value.

type constant = Ast.Term.constant

De Bruijn levels (not indexes): the distance of the binder from the root. starts at 0 and grows for bound variables; global constants have negative values.

type builtin =

| Cut

| And

| Impl

| ImplBang

| RImpl

| Pi

| Sigma

| Eq

| Match

| Findall

| Delay

| Host of constant

type term = Data.term

type view = private

| Const of int

| Lam of term

| App of int * term * term list

| Cons of term * term

| Nil

| Builtin of builtin * term list

| CData of RawOpaqueData.t

| UnifVar of FlexibleData.Elpi.t * term list

val look : depth:int -> term -> view

Terms must be inspected after dereferencing their head. If the resulting term is UVar then its uvar_body is such that get_assignment uvar_body = None

val kool : view -> term

val mkBound : int -> term

Smart constructors

val mkLam : term -> term

val mkCons : term -> term -> term

val mkNil : term

val mkDiscard : term

val mkCData : RawOpaqueData.t -> term

val mkUnifVar : FlexibleData.Elpi.t -> args:term list -> State.t -> term

val mkGlobal : constant -> term

Lower level smart constructors

val mkAppGlobal : constant -> term -> term list -> term

val mkAppGlobalL : constant -> term list -> term

val mkAppBound : int -> term -> term list -> term

val mkAppBoundL : int -> term list -> term

val mkBuiltin : builtin -> term list -> term

val mkConst : int -> term

no check, works for globals and bound

val mkApp : int -> term -> term list -> term

val mkAppMoreArgs : depth:int -> term -> term list -> term

val isApp : depth:int -> term -> bool

val cmp_builtin : builtin -> builtin -> int

type hyp = {

hdepth : int;

hsrc : term;

}

type hyps = hyp list

val of_hyp : Data.hyp -> hyp

val of_hyps : Data.hyp list -> hyps

type blockers = FlexibleData.Elpi.t list

type suspended_goal = {

context : hyps;

goal : int * term;

blockers : blockers;

}

val constraints : Data.constraints -> suspended_goal list

val no_constraints : Data.constraints

module Constants : sig ... end

type Conversion.extra_goal +=

| RawGoal of Data.term

val set_extra_goals_postprocessing : ?descriptor:Setup.hoas_descriptor -> (Conversion.extra_goals -> State.t -> State.t * Conversion.extra_goals) -> unit

val new_hoas_descriptor : unit -> Setup.hoas_descriptor