Module API.RawPp

val term : int -> Stdlib.Format.formatter -> Data.term -> unit

If the term is under depth binders this is the function that has to be * called in order to print the term correct. WARNING: as of today printing * an open term (i.e. containing unification variables) in the *wrong* depth * can cause the pruning of the unification variable. * This behavior shall be cleaned up in the future

val constraints : Stdlib.Format.formatter -> Data.constraints -> unit

val list : ?max:int -> ?boxed:bool -> (Stdlib.Format.formatter -> 'a -> unit) -> ?pplastelem:(Stdlib.Format.formatter -> 'a -> unit) -> string -> Stdlib.Format.formatter -> 'a list -> unit

module Debug : sig ... end