elpi index

Library elpi

The entry point of this library is the module: Elpi.

Library elpi.lexer_config

The entry point of this library is the module: Elpi_lexer_config.

Library elpi.parser

The entry point of this library is the module: Elpi_parser.

Library elpi.trace.ppx

The entry point of this library is the module: Trace_ppx.

Library elpi.trace.runtime

The entry point of this library is the module: Trace_ppx_runtime.

Library elpi.util

The entry point of this library is the module: Elpi_util.